Update contribution and developer guidance (#126)
### Change Description
<!-- Use this section to describe or list, at a high level, the changes
contained in the PR. Can be in a concise format as you would see on a
changelog. -->

This PR updates the contribution guidance for maintainers and community
contributors to the project.

There are updates to:
- The PR template
- The project design
- The testing strategy
- VSCode base settings (spelling, format, suggested extensions, tasks)

The intention is to streamline the PR creation and review processes,
potentially with automation. We would use the same template and approach
between all projects.

_Note to reviewers: If updating the test strategy or provider project
design wording, please update the files directly in a new PR to this
branch (or pull local and update this branch) rather than add comments_